requirements.txt
fastapi==0.95.0
pydantic==1.10.7
python-multipart==0.0.6
pymongo==3.13.0
bson==0.5.10
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1